pyproject.toml
[project]
name = "stock-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cffi>=2.0.0",
    "finnhub-python>=2.4.25",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-nvidia-ai-endpoints>=0.3.18",
    "matplotlib>=3.10.7",
    "openai>=2.4.0",
    "pandas>=2.3.3",
    "python-telegram-bot==20.*",
    "tabulate>=0.9.0",
    "yfinance>=0.2.66",
]
